fixed navigation for whole app, added theme for enitre app. added EsLint Rules and Prettier Formatting for project.
