+++ .github/pull_request_template.md
@@ -1,0 +1,13 @@
+## Vad har du gjort?
+
+en beskriv över vad din PR gör ....
+
+## Acceptanskrav
+
+- en användare ska kunna...
+- en ägare ska kunna...
+
+## Ska kontrolleras
+
+- [ ] Att temafärger används framför hårdkodade färger
+- [ ] Jag har kontrollerat att acceptanskraven fungerar
