+++ .github/workflows/workflows.yml
@@ -1,0 +1,23 @@
+name: CHECK PR CI
+
+on:
+  pull_request:
+    branches: ["main"]
+
+jobs:
+  check-pr:
+    runs-on: ubuntu-latest
+
+    strategy:
+      matrix:
+        node-version: [18.x]
+
+    steps:
+      - uses: actions/checkout@v3
+      - name: Use Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@v3
+        with:
+          node-version: ${{ matrix.node-version }}
+          cache: "npm"
+      - run: npm ci
+      - run: npm run types
