+++ package.json
@@ -6,7 +6,9 @@
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web"
+    "web": "expo start --web",
+    "types": "tsc --noEmit",
+    "lint": "eslint ./src"
   },
   "dependencies": {
     "@react-navigation/native-stack": "^6.9.14",
@@ -20,6 +22,10 @@
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.14",
+    "eslint": "^8.51.0",
+    "eslint-config-universe": "^12.0.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "^3.0.3",
     "typescript": "^5.1.3"
   },
   "private": true
