+++ src/Screen/CreateAccountScreen.tsx
@@ -1,15 +1,15 @@
-import React from "react";
-import { View, Text, StyleSheet, Image } from "react-native";
-import { NativeStackScreenProps } from "@react-navigation/native-stack";
-import { RootStackParamList } from "../Navigation/Navigation";
+import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
 
-type Props = NativeStackScreenProps<RootStackParamList, "CreateAccount">;
+import { RootStackParamList } from '../Navigation/RootNavigator';
+
+type Props = NativeStackScreenProps<RootStackParamList, 'CreateAccount'>;
 
 export default function CreateAccountScreen({ navigation }: Props) {
   return (
     <View style={styles.container}>
       <Text>Create Account Screen</Text>
-      
     </View>
   );
 }
@@ -17,8 +17,8 @@ export default function CreateAccountScreen({ navigation }: Props) {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    alignItems: "center",
-    justifyContent: "center", 
-    backgroundColor: "white",
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: 'white',
   },
 });
