+++ src/Screen/ProfileSettingScreen.tsx
@@ -1,15 +1,15 @@
-import React from "react";
-import { View, Text, StyleSheet, Image } from "react-native";
-import { NativeStackScreenProps } from "@react-navigation/native-stack";
-import { RootStackParamList } from "../Navigation/Navigation";
+import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
 
-type Props = NativeStackScreenProps<RootStackParamList, "ProfileSettings">;
+import { RootStackParamList } from '../Navigation/RootNavigator';
+
+type Props = NativeStackScreenProps<RootStackParamList, 'ProfileSettings'>;
 
 export default function ProfileSettingScreen({ navigation }: Props) {
   return (
     <View style={styles.container}>
       <Text>Profile Setting Screen</Text>
-      
     </View>
   );
 }
@@ -17,8 +17,8 @@ export default function ProfileSettingScreen({ navigation }: Props) {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    alignItems: "center",
-    justifyContent: "center", 
-    backgroundColor: "white",
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: 'white',
   },
 });
