+++ theme/theme.ts
@@ -3,11 +3,11 @@ export const ProjectTheme = {
   ...DefaultTheme,
   colors: {
     ...DefaultTheme.colors,
-    primary: "#FFFFFF",
-    secondary: "#0097a7",
-    itemcolor: "#FFFFFF",
+    primary: "#099E2C",
+    secondary: "#FFFFFF",
     iconcolor: "#000000",
     textcolor: "#333333",
+    whiteTextColor: "#FFFFFF",
     background: "#F2F2F2",
     warningHighLightIcon: "CD5D6F",
   },
@@ -53,11 +53,11 @@ export const ProjectTheme = {
     large: 8,
   },
   buttonPrimary: {
-    backgroundColor: "#0097a7",
-    color: "#333333",
+    backgroundColor: "#099E2C",
+    color: "#FFFFFF",
   },
   buttonSecondary: {
-    backgroundColor: "#0097a7",
+    backgroundColor: "#FFFFFF",
     color: "#333333",
   },
   inputBackground: "#FFFFFF",
@@ -73,7 +73,7 @@ export const ProjectTheme = {
   iconSize: 24,
   iconFamily: "MaterialIcons",
   dividerColor: "#E0E0E0",
-  headerBackgroundColor: "#F2F2F2",
+  headerBackgroundColor: "#FFFFFF",
   headerTitleColor: "#333333",
   navigationBackground: "#F2F2F2",
   modalBackground: "rgba(0, 0, 0, 0.8)",
